requirements.txt
streamlit
pypdf
sentence-transformers
chromadb
openai
python-dotenv